# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Recruitment ATS + CRM API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
...
  /users:
    get:
      operationId: listUsers
      tags: [auth]
      summary: List users
      parameters:
        - name: role
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [auth]
      summary: Create user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    patch:
      operationId: updateUser
      tags: [auth]
      summary: Update user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [auth]
      summary: Delete user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── CLIENTS ─────────────────────────────────────────────────────────────────
...
  /candidates:
    get:
      operationId: listCandidates
      tags: [candidates]
      summary: List candidates
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: positionId
          in: query
          schema:
            type: integer
        - name: clientId
          in: query
          schema:
            type: integer
        - name: source
          in: query
          schema:
            type: string
        - name: recruiterId
          in: query
          schema:
            type: integer
        - name: dateFrom
          in: query
          schema:
            type: string
        - name: dateTo
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Candidates list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Candidate"
    post:
      operationId: createCandidate
      tags: [candidates]
      summary: Submit candidate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CandidateInput"
      responses:
        "201":
          description: Submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Candidate"
  /candidates/{id}:
    get:
      operationId: getCandidate
      tags: [candidates]
      summary: Get candidate
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Candidate detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CandidateDetail"
    patch:
      operationId: updateCandidate
      tags: [candidates]
      summary: Update candidate
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CandidateUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Candidate"
    delete:
      operationId: deleteCandidate
      tags: [candidates]
      summary: Delete candidate
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── NOTES ────────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [user]
      properties:
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      required: [id, name, email, role, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
        createdAt:
          type: string
    UserInput:
      type: object
      required: [name, email, password, role]
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
    UserUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        role:
          type: string
        password:
          type: string

    # ─── CLIENT ────────────────────────────────────────────────────────────────
...
    Candidate:
      type: object
      required: [id, positionId, positionName, clientId, clientName, candidateName, email, phone, status, source, submissionDate, createdAt]
      properties:
        id:
          type: integer
        positionId:
          type: integer
        positionName:
          type: string
        clientId:
          type: integer
        clientName:
          type: string
        candidateName:
          type: string
        email:
          type: string
        phone:
          type: string
        currentCompany:
          type: ["string", "null"]
        experience:
          type: ["string", "null"]
        currentCtc:
          type: ["string", "null"]
        expectedCtc:
          type: ["string", "null"]
        noticePeriod:
          type: ["string", "null"]
        source:
          type: string
        status:
          type: string
        submissionDate:
          type: string
        createdAt:
          type: string
        recruiterId:
          type: ["integer", "null"]
        recruiterName:
          type: ["string", "null"]
        resumeFileName:
          type: ["string", "null"]
    CandidateDetail:
      type: object
      required: [id, positionId, positionName, clientId, clientName, candidateName, email, phone, status, source, submissionDate, createdAt]
      properties:
        id:
          type: integer
        positionId:
          type: integer
        positionName:
          type: string
        clientId:
          type: integer
        clientName:
          type: string
        candidateName:
          type: string
        email:
          type: string
        phone:
          type: string
        currentCompany:
          type: ["string", "null"]
        experience:
          type: ["string", "null"]
        currentCtc:
          type: ["string", "null"]
        expectedCtc:
          type: ["string", "null"]
        noticePeriod:
          type: ["string", "null"]
        source:
          type: string
        status:
          type: string
        submissionDate:
          type: string
        createdAt:
          type: string
        recruiterId:
          type: ["integer", "null"]
        recruiterName:
          type: ["string", "null"]
        resumeFileName:
          type: ["string", "null"]
    CandidateInput:
      type: object
      required: [positionId, candidateName, email, phone, source]
      properties:
        positionId:
          type: integer
        candidateName:
          type: string
        email:
          type: string
        phone:
          type: string
        currentCompany:
          type: string
        experience:
          type: string
        currentCtc:
          type: string
        expectedCtc:
          type: string
        noticePeriod:
          type: string
        source:
          type: string
        submissionDate:
          type: string
        resumeFileName:
          type: string
        resumeData:
          type: string
    CandidateUpdate:
      type: object
      properties:
        candidateName:
          type: string
        email:
          type: string
        phone:
          type: string
        currentCompany:
          type: string
        experience:
          type: string
        currentCtc:
          type: string
        expectedCtc:
          type: string
        noticePeriod:
          type: string
        source:
          type: string
        status:
          type: string
        resumeFileName:
          type: string
        resumeData:
          type: string

    # ─── NOTES ─────────────────────────────────────────────────────────────────
...