# lib/api-zod/src/generated/types/userUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Recruitment ATS + CRM API
 * OpenAPI spec version: 0.1.0
 */

export interface UserUpdate {
  name?: string;
  email?: string;
  role?: string;
  password?: string;
}